refactor: 패키지상세화 및 클래스명 변경
패키지 상세화:
- domain, view, util 패키지 생성

클래스명 변경:
- Playground -> GameController로 변경
