+++ src/main/java/edu/nextstep/racing/RacingMain.java
@@ -3,6 +3,6 @@ package edu.nextstep.racing;
 public class RacingMain {
 
 	public static void main(String[] args) {
-		new Playground().play();
+		new GameController().play();
 	}
 }
