+++ src/main/java/edu/nextstep/racing/view/ResultView.java
@@ -1,6 +1,8 @@
-package edu.nextstep.racing;
+package edu.nextstep.racing.view;
 
 import java.util.List;
+
+import edu.nextstep.racing.domain.Car;
 
 public class ResultView {
